[modify] 어드민 당첨자 모달 전화번호에 하이픈 추가
