+++ src/common/components/PhoneInput.jsx
@@ -1,14 +1,7 @@
 import Input from "./Input.jsx";
+import { addHyphen } from "../utils.js";
 
 function PhoneInput({ text, setText, ...otherProps }) {
-  function addHyphen(value) {
-    const plain = value.replace(/\D/g, "");
-
-    if (plain.length < 4) return plain;
-    if (plain.length <= 7) return plain.replace(/^(\d{3})(\d{0,4})$/, "$1-$2");
-    if (plain.length <= 10) return plain.replace(/^(\d{3})(\d{3})(\d{0,4})$/, "$1-$2-$3");
-    return plain.replace(/^(\d{3})(\d{4})(\d{4,})$/, "$1-$2-$3");
-  }
   return (
     <Input
       text={text}
