+++ src/common/utils.js
@@ -145,3 +145,12 @@ export function getDayDifference(_date1, _date2) {
 
   return dayDifference;
 }
+
+export function addHyphen(value) {
+  const plain = value.replace(/\D/g, "");
+
+  if (plain.length < 4) return plain;
+  if (plain.length <= 7) return plain.replace(/^(\d{3})(\d{0,4})$/, "$1-$2");
+  if (plain.length <= 10) return plain.replace(/^(\d{3})(\d{3})(\d{0,4})$/, "$1-$2-$3");
+  return plain.replace(/^(\d{3})(\d{4})(\d{4,})$/, "$1-$2-$3");
+}
